[sc-931]Discard Studies & Decrease amount
